spv ffi wrappers reorganized and SPVClient made sendable
